+++ packages/expo-widgets/src/ExpoWidgets.ios.ts
@@ -5,10 +5,15 @@ import type {
   NativeLiveActivity,
   NativeLiveActivityFactory,
   NativeWidgetObject,
+  WidgetImageClearOptions,
+  WidgetImagePreloadOptions,
+  WidgetImagePreloadResult,
 } from './Widgets.types';
 
 declare class ExpoWidgetsModule extends NativeModule<ExpoWidgetsEvents> {
   reloadAllWidgets(): void;
+  preloadImagesAsync(images: WidgetImagePreloadOptions[]): Promise<WidgetImagePreloadResult>;
+  clearPreloadedImagesAsync(options?: WidgetImageClearOptions): Promise<void>;
   readonly Widget: typeof NativeWidgetObject;
   readonly LiveActivityFactory: typeof NativeLiveActivityFactory;
   readonly LiveActivity: typeof NativeLiveActivity;