+++ packages/expo-widgets/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - [Android] Create base android package. ([#46090](https://github.com/expo/expo/pull/46090) by [@jakex7](https://github.com/jakex7))
 - [plugin] Create config plugin for android. ([#46091](https://github.com/expo/expo/pull/46091) by [@jakex7](https://github.com/jakex7))
+- Add APIs for preloading remote images into widget-accessible app group storage.
 
 ### 🐛 Bug fixes